Fix clear highlighter selection when using multiselect (#590)
* Fix: issue that clears the highlighter while using multiselect

When using multiselect (in this example ctrl-key) and selecting multiple elements, a click that does not select any new object (aka clicking anywhere except on a element) will clear the selection, which is not intuitive and NOT the behavior for any 3D modeling software.

I'd like to get this to the 2.3.x builds but id guess this is not possible correct?

* Fix: issue that clears the highlighter while using multiselect

When using multiselect (in this example ctrl-key) and selecting multiple elements, a click that does not select any new object (aka clicking anywhere except on a element) will clear the selection, which is not intuitive and NOT the behavior for any 3D modeling software.

I'd like to get this to the 2.3.x builds but id guess this is not possible correct?